backend/requirements.txt
fastapi
uvicorn
pandas
gspread
google-auth
supabase
python-dotenv
toml
pyjwt
xgboost
scikit-learn
requests
pytesseract
pillow
